Adding Resume to Download
Able to download and view at the same time but not view only so it is still not fully done, I also am not going to do much with it until I finish with the other sections. Getting the layout of the front end done would probably be very beneficial and reduce the amount of time it takes to do things.