Reformat/Refactor Files
I decided to refactor my code to make it the editing/reusability aspect of it a lot easier. I also asked for some help on changing the layout of the resume as well. I am studying the code to learn and understand what was done (:, I also realized the box shadow is now popping out a lot less so I will take a look at that and try to work through it!