Implemented Skills Carousel
This took yesterday and today to get done. Still not completely done with it, but happy with the way it is made. The backend implementation for this website is about to begin soon. Final exams are swooping in so I need to make sure to budget some time for school before continuing all my time into this project.